+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Features
 * **@dlr-eoc/map-ol:**
  - The function `zoomInOut` now accepts an optional object with `callback`, `duration`, `easing` and `zoomStep`.
-
+ - Add types for `FitOptions` on `setExtent`.
 ### Bug Fixes
 * **@dlr-eoc/map-ol:**
  - Remove LayerGroups from map before add `baselayerGroup`, `layersGroup` and `overlayGroup`.